fix: move around some pages for consistency
