+++ frontend/lib/navigation/app_navigation.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:nibbles/pages/profile_page.dart';
-import '../pages/home_page.dart';
+import 'package:nibbles/pages/profile/profile_page.dart';
+import '../pages/home/home_page.dart';
 import '../pages/recipes/recipes_page.dart';
-import '../pages/map_page.dart';
+import '../pages/map/map_page.dart';
 
 class AppNavigation extends StatefulWidget {
   const AppNavigation({super.key});
