--- frontend/lib/pages/restaurants/restaurants_page.dart
@@ -1,19 +1,0 @@
-import 'package:flutter/material.dart';
-
-class RestaurantsPage extends StatelessWidget {
-  const RestaurantsPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: SafeArea(
-        child: SingleChildScrollView(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [Center(child: Text("Restaurants"))],
-          ),
-        ),
-      ),
-    );
-  }
-}
